remove all funtion duplicate in mailbox_dashboard_controller
(cherry picked from commit 2cbe5f163c3478d57d0db69dac0e2aa23a486006)
